+++ core/src/com/stewsters/fighter/actor/pilot/HumanPilot.kt
@@ -1,0 +1,31 @@
+package com.stewsters.fighter.actor.pilot
+
+import com.badlogic.gdx.controllers.Controller
+import com.badlogic.gdx.controllers.mappings.Xbox
+import com.stewsters.fighter.FighterGame
+import com.stewsters.fighter.actor.Actor
+import com.stewsters.fighter.math.accel
+import com.stewsters.fighter.math.deaden
+
+class HumanPilot(val controller: Controller) : PilotBase() {
+    override fun fly(fighterGame: FighterGame, us: Actor) {
+        with(us) {
+            if (aircraftType != null) {
+
+                accelp = aircraftType.acceleration * accel(-controller.getAxis(Xbox.L_STICK_VERTICAL_AXIS))
+                yawp = aircraftType.turnYaw * deaden(controller.getAxis(Xbox.L_STICK_HORIZONTAL_AXIS))
+                pitchp = aircraftType.turn * deaden(controller.getAxis(Xbox.R_STICK_VERTICAL_AXIS))
+                rollp = aircraftType.turn * deaden(controller.getAxis(Xbox.R_STICK_HORIZONTAL_AXIS))
+
+                if (primaryWeapon != null && controller.getButton(Xbox.R_BUMPER)) {
+                    primaryWeapon.fire(fighterGame, this)
+                }
+
+                if (secondaryWeapon != null && controller.getButton(Xbox.L_BUMPER)) {
+                    secondaryWeapon.fire(fighterGame, this)
+                }
+
+            }
+        }
+    }
+}
