+++ core/src/com/stewsters/fighter/types/Types.kt
@@ -1,0 +1,11 @@
+package com.stewsters.fighter.types
+
+import com.badlogic.gdx.graphics.VertexAttributes
+import com.badlogic.gdx.graphics.g3d.loader.ObjLoader
+import com.badlogic.gdx.graphics.g3d.utils.ModelBuilder
+
+val attr = (VertexAttributes.Usage.Position or VertexAttributes.Usage.Normal).toLong()
+val modelBuilder = ModelBuilder()
+val loader = ObjLoader()
+
+
